Add hamburger menu for small screens, refactor navbar
